refactor(hickey): extract shared SubCountChip
Both the desktop dock and the mobile dock row were carrying a
byte-equivalent chip component (SubCountChip + MobileSubCountChip)
— same Tailwind classes, same active/inactive color math, same
icon, same label accessor. Concept multiplication for a single
volatility axis (how a sub-terminal count renders).

Extracts a shared SubCountChip to packages/client/src/canvas/dock/
with the surface area the call sites actually need: `count`,
`active`, and a required `testId` so each surface stays
distinguishable in tests without a default-value convention. The
optional `class` prop is dropped — caller positioning belongs on
the wrapping element, which the previous commit already
restructured in QuietRowBody.
